Split utils.py into apply.py and born.py.
