+++ bgls/utils.py
@@ -21,116 +21,6 @@ import cirq
 import bgls
 
 
-def cirq_state_vector_bitstring_probability(
-    state_vector_state: cirq.sim.state_vector_simulation_state.StateVectorSimulationState,
-    bitstring: str,
-) -> float:
-    """Returns the probability of measuring the `bitstring` (|z⟩) in the
-    `cirq.StateVectorSimulationState` (|ψ⟩), i.e. |⟨z|ψ⟩|^2.
-
-    Args:
-        state_vector_state: State vector |ψ⟩ as a
-            `cirq.StateVectorSimulationState`.
-        bitstring: Bitstring |z⟩ as a binary string.
-    """
-    num_qubits = len(state_vector_state.qubits)
-    return (
-        np.abs(
-            cirq.to_valid_state_vector(
-                state_vector_state.target_tensor, num_qubits=num_qubits
-            )[int(bitstring, 2)]
-        )
-        ** 2
-    )
-
-
-def cirq_density_matrix_bitstring_probability(
-    density_matrix_state: cirq.sim.DensityMatrixSimulationState,
-    bitstring: str,
-) -> float:
-    """Returns the probability of measuring the `bitstring` (|z⟩) in the
-    `cirq.DensityMatrixSimulationState` (ρ), i.e. ⟨z|ρ|z⟩.
-
-    Args:
-        density_matrix_state: Density matrix ρ as a
-            `cirq.DensityMatrixSimulationState`.
-        bitstring: Bitstring |z⟩ as a binary string.
-    """
-    num_qubits = len(bitstring)
-    index = int(bitstring, 2)
-    density_matrix = cirq.to_valid_density_matrix(
-        density_matrix_state.target_tensor, num_qubits
-    )
-    return np.abs(density_matrix[index, index])
-
-
-def cirq_stabilizer_ch_bitstring_probability(
-    stabilizer_ch_form_state: cirq.sim.StabilizerChFormSimulationState,
-    bitstring: str,
-) -> float:
-    """Returns the probability of measuring the `bitstring` (|z⟩) in the
-    `cirq.StabilizerChFormSimulationState` (U_C U_H|s⟩), i.e. |⟨z|ψ⟩|^2.
-
-    Args:
-        stabilizer_ch_form_state: Stabilizer state in CH form (arxiv:
-        1808.00128) as a 'cirq.StabilizerChFormSimulationState'.
-        bitstring: Bitstring |z⟩ as a binary string.
-
-    """
-    # the state is of type StabilizerStateChForm
-    # this runs in O(n^2) for an n qubit state
-    return (
-        np.abs(
-            stabilizer_ch_form_state.state.inner_product_of_state_and_x(
-                int(bitstring, 2)
-            )
-        )
-        ** 2
-    )
-
-
-def act_on_near_clifford(
-    op: cirq.Operation,
-    state: bgls.simulator.State,
-    rng: np.random.RandomState = np.random.RandomState(),
-) -> None:
-    """
-    Applies a Clifford+T (or more generically, Rz(theta)) operation to a state.
-    If the operation is Clifford, apply normally,
-    else choose one of the operations in the Clifford-expansion to apply.
-
-    Parameters:
-        op: operation to apply.
-        state: simulator state to update.
-        rng: rng for controlling gate-expansion decisions.
-    """
-    if cirq.has_stabilizer_effect(op):
-        cirq.protocols.act_on(op, state)
-    elif isinstance(op.gate, cirq.ops.common_gates.ZPowGate):
-        # note this includes Rz gates
-        theta = np.pi * op.gate.exponent
-
-        probs = np.power(
-            np.abs(
-                [
-                    np.cos(theta / 2) - np.sin(theta / 2),
-                    np.sqrt(2.0)
-                    * np.exp(-(0 + 1j) * np.pi / 4)
-                    * np.sin(theta / 2),
-                ]
-            ),
-            2,
-        )
-        cand_gates = [cirq.I(op.qubits[0]), cirq.S(op.qubits[0])]
-        chosen_gate = rng.choice(cand_gates, p=probs / sum(probs))
-        cirq.protocols.act_on(chosen_gate, state)
-    else:
-        raise ValueError(
-            "Operation must be Clifford, T, or a z rotation. Was called on "
-            "the gate " + str(op.gate) + "."
-        )
-
-
 def generate_random_circuit(
     qubits: Union[Sequence[cirq.ops.Qid], int],
     n_moments: int,
