+++ bgls/utils_test.py
@@ -13,7 +13,6 @@
 """Tests for the BGLS Simulator utilities."""
 
 import pytest
-import numpy as np
 
 import cirq
 
@@ -41,29 +40,3 @@ def test_generate_random_circuit_matches_cirq_random_circuit(nqubits: int):
         random_state=1,
     )
     assert clifford_circuit2 == clifford_circuit
-
-
-def test_act_on_stabilizer_non_T():
-    """Assertion thrown if used on a non-T non-Rz non-Clifford gate"""
-    qubits = cirq.LineQubit.range(3)
-    state = cirq.StateVectorSimulationState(qubits=qubits, initial_state=0)
-    # fails for non zpowgate
-    with pytest.raises(ValueError):
-        gate = cirq.XPowGate(exponent=0.1)
-        bgls.utils.act_on_near_clifford(gate.on(qubits[0]), state)
-
-
-@pytest.mark.parametrize("rads", [0.0, 0.5, 1.0, 1.5, np.pi + 0.1])
-def test_generic_rz_gates(rads: float):
-    """act_on_near_clifford supports other rz rotations besides T (pi/4)"""
-    qubits = cirq.LineQubit.range(3)
-    state1 = cirq.StateVectorSimulationState(qubits=qubits, initial_state=0)
-    state2 = cirq.StateVectorSimulationState(qubits=qubits, initial_state=0)
-    state2S = state2.copy()
-    cirq.act_on(cirq.S.on(qubits[0]), state2S)
-
-    op = cirq.Rz(rads=rads)
-    bgls.utils.act_on_near_clifford(op.on(qubits[0]), state1)
-    assert (state1.target_tensor == state2.target_tensor).all() or (
-        state1.target_tensor == state2S.target_tensor
-    ).all()
